refactor: preliminary tidies to the whole conformance test suite
